[wip] trying to get the latest version of Typedoc working
Getting an error where the types for 'node-fetch' cannot be found, but
nothing that I can do from the config to ignore `node_modules` or `http`
is working, even though both of those things are ignored as specified in
the config. Will have to deep-dive into typedoc itself and find out why
it's not working, or find an alternative.